refactor: 회원가입 이벤트 리스너 추가
