+++ src/main/java/com/runky/auth/application/AuthFacade.java
@@ -11,16 +11,10 @@ import com.runky.auth.domain.port.TokenDecoder;
 import com.runky.auth.domain.signup.SignupTokenService;
 import com.runky.auth.domain.vo.OAuthUserInfo;
 import com.runky.auth.domain.vo.RefreshTokenClaims;
-import com.runky.crew.domain.CrewCommand;
-import com.runky.crew.domain.CrewService;
-import com.runky.goal.domain.GoalCommand;
-import com.runky.goal.domain.GoalService;
 import com.runky.member.domain.MemberCommand;
 import com.runky.member.domain.dto.MemberInfo;
 import com.runky.member.domain.service.MemberReader;
 import com.runky.member.domain.service.MemberRegistrar;
-import com.runky.reward.domain.RewardCommand;
-import com.runky.reward.domain.RewardService;
 
 import lombok.RequiredArgsConstructor;
 
@@ -36,9 +30,6 @@ public class AuthFacade {
 
 	private final TokenDecoder tokenDecoder;
 	private final AuthTokenService authTokenService;
-	private final RewardService rewardService;
-	private final GoalService goalService;
-	private final CrewService crewService;
 
 	private final ApplicationEventPublisher eventPublisher;
 
@@ -79,10 +70,7 @@ public class AuthFacade {
 			));
 		AuthInfo.TokenPair pair = authTokenService.issue(saved.id(), saved.role().name());
 		signupTokenService.delete(signupToken);
-		rewardService.init(new RewardCommand.Init(saved.id()));
-		goalService.init(new GoalCommand.Init(saved.id()));
-		crewService.init(new CrewCommand.Init(saved.id()));
-		//		eventPublisher.publishEvent(new AuthEvent.SignupCompleted(saved.id()));
+        eventPublisher.publishEvent(new AuthEvent.SignupCompleted(saved.id()));
 
 		return new AuthResult.SigninComplete(pair.accessToken(), pair.refreshToken());
 	}
@@ -98,9 +86,5 @@ public class AuthFacade {
 	public void logoutByRefresh(String refreshToken) {
 		RefreshTokenClaims decoded = tokenDecoder.decodeRefresh(refreshToken);
 		authTokenService.delete(decoded.memberId());
-	}
-
-	public void deleteAccountByRefresh(final String refreshToken) {
-
 	}
 }
