+++ src/main/java/com/runky/crew/interfaces/api/CrewApiSpec.java
@@ -1,4 +1,4 @@
-package com.runky.crew.interfaces;
+package com.runky.crew.interfaces.api;
 
 import com.runky.global.response.ApiResponse;
 import com.runky.global.security.auth.MemberPrincipal;
