+++ src/main/java/com/runky/reward/interfaces/api/RewardApiSpec.java
@@ -1,9 +1,9 @@
-package com.runky.reward.interfaces;
+package com.runky.reward.interfaces.api;
 
 import com.runky.global.response.ApiResponse;
 import com.runky.global.security.auth.MemberPrincipal;
-import com.runky.reward.interfaces.RewardResponse.Gotcha;
-import com.runky.reward.interfaces.RewardResponse.Images;
+import com.runky.reward.interfaces.api.RewardResponse.Gotcha;
+import com.runky.reward.interfaces.api.RewardResponse.Images;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.Parameter;
