+++ src/main/java/com/runky/reward/interfaces/api/RewardResponse.java
@@ -1,4 +1,4 @@
-package com.runky.reward.interfaces;
+package com.runky.reward.interfaces.api;
 
 import java.util.List;
